+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Examples
 ........
 
   * `apprtc` : make the initiator send messages using an HTTP POST instead of WebSocket.
+  * `server` : add cartoon effect to video transforms.
 
 0.9.18
 ------